requirements.txt
amqp==5.0.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap3==21.2
django-braces==1.15.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-debug-toolbar==3.2.4
django-timezone-field==4.2.3
djangorestframework==3.13.1
Faker==11.1.0
gunicorn==20.1.0
kombu==5.2.3
misaka==2.1.1
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.28
psycopg2==2.8.6
pycparser==2.21
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.1.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.2
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
